Tests topics

+++ src/controllers/topicsController.js
@@ -1,8 +1,12 @@
 const Topic = require('../models/Topic');
 const InexistingId = require('../errors/InexistingId');
+const Chapter = require('../models/Chapter');
 
 class TopicsController {
   async createTopic({ chapterId, name }) {
+    const chapter = await Chapter.findByPk(chapterId);
+    if (!chapter) throw new InexistingId();
+
     const topic = await Topic.create({ chapterId, name });
     return topic;
   }
@@ -26,11 +30,6 @@ class TopicsController {
     return topics;
   }
 
-  async getAllTopicsAsAdmin() {
-    const topics = await Topic.findAll();
-    return topics;
-  }
-
   async getTopicByIdAsAdmin(id) {
     const topic = await Topic.findByPk(id);
     if (!topic) throw new InexistingId();
@@ -39,6 +38,8 @@ class TopicsController {
 
   async editTopic(id, { name, chapterId }) {
     const topic = await Topic.findByPk(id);
+    if (!topic) throw new InexistingId();
+
     topic.name = name || topic.name;
     topic.chapterId = chapterId || topic.chapterId;
     await topic.save();
@@ -46,6 +47,9 @@ class TopicsController {
   }
 
   async deleteOneTopic(id) {
+    const topic = await Topic.findByPk(id);
+    if (!topic) throw new InexistingId();
+
     await Topic.destroy({
       where: { id },
     });
